pyproject.toml
[project]
name = "nebus-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Twopercent",email = "twopercent051@zohomail.eu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "betterlogging (>=1.0.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.34.2,<0.35.0)",

]


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
pytest = "^8.3.5"

